# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peres Bakeries Limited API v1",
        "description": "Peres Bakeries Limited API for managing daily transactons in bakery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/v1/staff/logout": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Deauthenticate a staff",
                "operationId": "Stafflogout",
                "responses": {
                    "200": {
                        "description": "Successful logout operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Successfully logged out"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "staff/refresh": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Refresh staff token after expiration",
                "operationId": "staffrefresh",
                "responses": {
                    "200": {
                        "description": "Successful token refresh",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "token": "string",
                                            "user": {
                                                "surname": "string",
                                                "firstname": "string",
                                                "gender": "string",
                                                "phone_number": "number",
                                                "address": "string",
                                                "position": "string",
                                                "dob": "string"
                                            },
                                            "token_type": "bearer",
                                            "expires_in": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "staff/firstlogin": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Login staff token after expiration",
                "operationId": "staffFirstLogin",
                "responses": {
                    "200": {
                        "description": "Successful first login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "0": "login_type':'first",
                                            "token": "string",
                                            "user": {
                                                "surname": "string",
                                                "firstname": "string",
                                                "gender": "string",
                                                "phone_number": "number",
                                                "address": "string",
                                                "position": "string",
                                                "dob": "string"
                                            },
                                            "token_type": "bearer",
                                            "expires_in": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/staff/login": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Staff login",
                "operationId": "stafflogin",
                "requestBody": {
                    "description": "Staf login object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StaffLoginModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "login_type": "string",
                                        "token": "string",
                                        "admin": {
                                            "id": "string",
                                            "surname": "string",
                                            "firstname": "string",
                                            "mobile_number": "number",
                                            "position": "string",
                                            "address": "string",
                                            "gender": "string",
                                            "dob": "string",
                                            "creared_at": "string"
                                        },
                                        "token_type": "bearer",
                                        "expires_in": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "error": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/v1/staff/cart": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create new cart item",
                "operationId": "AddToCart",
                "requestBody": {
                    "description": "Create cart object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCartModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item Successfully added"
                    },
                    "422": {
                        "description": "Unproccessable data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "error": "Unproccessable data"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/staff/cart/delete/{id}": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Delete a cart from storage",
                "operationId": "CartDelete",
                "responses": {
                    "200": {
                        "description": "Successful delete operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Cart successfully deleted"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateCartModel": {
                "title": "CreateProductModel",
                "description": "Pares Bakeries Limited Login Model",
                "properties": {
                    "product_id": {
                        "title": "Product Id",
                        "type": "string",
                        "example": "xxxxxxxxxxxxxxxxx"
                    },
                    "price": {
                        "title": "Product unit price",
                        "type": "string",
                        "example": "200"
                    },
                    "quantity": {
                        "title": "Product quantity",
                        "type": "string",
                        "example": "4"
                    },
                    "amount": {
                        "title": "Amount for Total Units",
                        "type": "string",
                        "example": 800
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CreateProductModel"
                }
            },
...
            "StaffLoginModel": {
                "title": "StudentLoginModel",
                "description": "Peres Bakeries sales management Staff Login Model",
                "properties": {
                    "mobile_number": {
                        "title": "Staff Phone Number",
                        "type": "string",
                        "example": "08012345678"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StaffLoginModel"
                }
            }
        }
    },
...
}